+++ NEWS.md
@@ -1,7 +1,8 @@
 # ggplot2 (development version)
 
-* `geom_text()` and `geom_label()` gained a `size.unit` parameter that can
-  switch between millimetres and points for the text size (@teunbrand, #3799).
+* `geom_text()` and `geom_label()` gained a `size.unit` parameter that set the 
+  text size to millimetres, points, centimetres, inches or picas 
+  (@teunbrand, #3799).
 * To improve `width` calculation in bar plots with empty factor levels, 
   `resolution()` considers `mapped_discrete` values as having resolution 1 
   (@teunbrand, #5211)